Changed type of 'Travel Amount' field to Big Decimal
Issue :- Type of 'Travel Amount' field was set as Integer in VO classes. Hence saving with higher amount was breaking the code

Solution :-  Changed type of 'Travel Amount' field to Big Decimal

Reviewer :- @lijisaji162

+++ Server/FCOI/src/main/java/com/polus/fibicomp/coi/dao/ConflictOfInterestDaoImpl.java
@@ -1008,7 +1008,7 @@ public class ConflictOfInterestDaoImpl implements ConflictOfInterestDao {
 					disclosureView.setTravelDisclosureStatusDescription(resultSet.getString("TRAVEL_DISCLOSURE_STATUS_DESCRIPTION"));
 					disclosureView.setTravelEntityName(resultSet.getString("TRAVEL_ENTITY_NAME"));
 					disclosureView.setTravellerName(resultSet.getString("TRAVELLER_NAME"));
-					disclosureView.setTravelAmount(resultSet.getInt("TRAVEL_AMOUNT"));
+					disclosureView.setTravelAmount(resultSet.getBigDecimal("TRAVEL_AMOUNT"));
 					disclosureView.setTravelSubmissionDate(resultSet.getDate("SUBMISSION_DATE"));
 					disclosureView.setAcknowledgeDate(resultSet.getDate("ACKNOWLEDGE_DATE"));
 					disclosureView.setTravelExpirationDate(resultSet.getDate("EXPIRATION_DATE"));
